+++ host-configs/lc-builds/toss3/oneapi_X.cmake
@@ -1,0 +1,16 @@
+###############################################################################
+# Copyright (c) 2016-22, Lawrence Livermore National Security, LLC
+# and RAJA project contributors. See the RAJA/LICENSE file for details.
+#
+# SPDX-License-Identifier: (BSD-3-Clause)
+###############################################################################
+
+set(CMAKE_CXX_FLAGS_RELEASE "${COMMON_FLAGS} -O3 -march=native -finline-functions" CACHE STRING "")
+set(CMAKE_CXX_FLAGS_RELWITHDEBINFO "${COMMON_FLAGS} -O3 -g -march=native -finline-functions" CACHE STRING "")
+set(CMAKE_CXX_FLAGS_DEBUG "${COMMON_FLAGS} -O0 -g" CACHE STRING "")
+
+set(RAJA_RANGE_ALIGN 4 CACHE INT "")
+set(RAJA_RANGE_MIN_LENGTH 32 CACHE INT "")
+set(RAJA_DATA_ALIGN 64 CACHE INT "")
+
+set(RAJA_HOST_CONFIG_LOADED On CACHE BOOL "")