+++ include/RAJA/policy/PolicyBase.hpp
@@ -49,7 +49,8 @@ enum class Pattern {
   workgroup,
   workgroup_exec,
   workgroup_order,
-  workgroup_storage
+  workgroup_storage,
+  workgroup_dispatch
 };
 
 enum class Launch { undefined, sync, async };